+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### Bugs fixed
 
+* Fix Style/ElseAlignment Cop to find the right parent on def/rescue/else/ensure/end. ([@oneamtu][])
 * [#1181](https://github.com/bbatsov/rubocop/issues/1181): *(fix again)* `Style/StringLiterals` cop stays away from strings inside interpolated expressions. ([@jonas054][])
 * [#1441](https://github.com/bbatsov/rubocop/issues/1441): Correct the logic used by `Style/Blocks` and other cops to determine if an auto-correction would alter the meaning of the code. ([@jonas054][])
 * [#1449](https://github.com/bbatsov/rubocop/issues/1449): Handle the case in `MultilineOperationIndentation` where instances of both correct style and unrecognized (plain wrong) style are detected during an `--auto-gen-config` run. ([@jonas054][])
@@ -1163,3 +1164,4 @@
 [@jfelchner]: https://github.com/jfelchner
 [@janraasch]: https://github.com/janraasch
 [@jcarbo]: https://github.com/jcarbo
+[@oneamtu]: https://github.com/oneamtu